requirements.txt
asgiref==3.2.8
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-heroku==0.3.1
dnspython==1.16.0
email-validator==1.1.1
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
